% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAS.R
\name{mice.impute.QAS}
\alias{mice.impute.QAS}
\title{Function to use QAS in Mice}
\usage{
mice.impute.QAS(y, ry, x)
}
\arguments{
\item{y}{The target variable}

\item{ry}{a logical vector, which indicates the missing values in the target variable}

\item{x}{a data frame with the independent variables}
}
\value{
An object of class \emph{mice.impute.QAS} is a vector containing the imputed values of the target variable.
}
\description{
\code{mice.impute.QAS} is used to prepare the \code{predictQAS} output for mice
}
\examples{
# generate Data
y <- as.integer(c(1,0,0,0,1,1,1,0,0,1))
x <- c(15,88,90,60,24,30,26,57,69,18)
z <- as.integer(c(3,2,2,1,3,3,2,1,1,3))
example_data <- data.frame(y,x,z)

# deploy QAS.func-Function
result1 <- QAS.func(y~x+z, data=example_data,weight=NULL,seed=NULL)

# deploy predictQas-Function
result2  <- predictQAS(QAS.res = result1, data=example_data)

# generate logical vector and data frame with independent variables
ry <- c(TRUE,TRUE,TRUE,FALSE,FALSE,TRUE,FALSE,TRUE,TRUE)
x <- data.frame(x,z)

# deploy mice.impute.QAS-Funktion
impute <- mice.impute.QAS(y=y,ry=ry,x=example_data)

}
