% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAS.R
\name{predictQAS}
\alias{predictQAS}
\title{Function for categorization after QAS.func}
\usage{
predictQAS(QAS.res, data)
}
\arguments{
\item{QAS.res}{The results of the QAS.func-Function}

\item{data}{A data frame containing the variables in the model for QAS.func}
}
\value{
An object of class \emph{predictQAS} is a list containing at least the following components:

\describe{
 \item{\code{yhat_orig}}{the predicted values of the dependent variable based on the coefficients of QAS.func and the dataset used for categorization}
 \item{\code{data}}{the dataset with categorized numeric variables}
 }
}
\description{
\code{predictQAS} is used to categorize  numeric variables in a dataset the same way as the QAS.func-Function for calculating the coefficients.
}
\examples{
# generate Data
y <- as.integer(c(1,0,0,0,1,1,1,0,0,1))
x <- c(15,88,90,60,24,30,26,57,69,18)
z <- as.integer(c(3,2,2,1,3,3,2,1,1,3))
example_data <- data.frame(y,x,z)

# deploy QAS.func-Function
result1 <- QAS.func(y~x+z, data=example_data,weight=NULL,seed=NULL)

# deploy predictQas-Function
result2  <- predictQAS(QAS.res = result1, data=example_data)


}
