% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAS.R
\name{QAS.func}
\alias{QAS.func}
\title{Quasi analytical solution for logit.}
\usage{
QAS.func(frml, data = NULL, weight = NULL, seed = NULL)
}
\arguments{
\item{frml}{an object of class \code{\link[stats]{formula}} (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.}

\item{data}{a data frame containing the variables in the model (or object coercible by \code{\link[base]{as.data.frame}} to a data frame).  Details of the structure of the data are given under 'Details'.}

\item{weight}{an optional vector of ‘prior weights’ to be used in the fitting process. Should be NULL or a numeric vector. In case of NULL, each case is weighted with 1.}

\item{seed}{saving the state of a random process. Should be NULL or a numeric vector. In case of NULL a seed is generated at random.}
}
\value{
An object of class \emph{QAS.func} is a list containing at least the following components:

\describe{
 \item{\code{coefficients}}{a vector of coefficients}
 \item{\code{weights}}{the working weights}
 \item{\code{call}}{the call of the final function within QAS.func}
 \item{\code{terms}}{the term object used}
 \item{\code{model}}{the model frame}
 \item{\code{y_hat}}{the predicted y values}
 \item{\code{means_for_cat}}{the cut points of the metric variables for a categorization of the original dataset}
 \item{\code{seed}}{used seed for calculations}
 \item{\code{AIC}}{A version of Akaike's Information Criterion}
 \item{\code{BIC}}{A version if Bayesian Information Criterion}
}
}
\description{
\code{QAS.func} is used to replace numerical optimization with a quasi-analytical approach for logit models on big data. It returns the coefficients, predicted values and quality criteria for the provided variables.
}
\details{
A typical predictor has the form dependent_Variable \strong{'~'} independent_Variables.\cr The dependent_Variable has \strong{two} categories.\cr If there is more than one independent_Variable, they can be combined with a '+'.

The data frame can \strong{not} contain any missing values.\cr \strong{Metric} variables have to be of type \strong{numeric}. All \strong{other variables} have to be of type \strong{integer}.\cr The first variable in the dataset hat to be the dependent variable.\cr The scale of large numbers has to be reduced e.g. standardization.
}
\examples{
# generate Data
y <- as.integer(c(1,0,0,0,1,1,1,0,0,1))
x <- c(15,88,90,60,24,30,26,57,69,18)
z <- as.integer(c(3,2,2,1,3,3,2,1,1,3))
example_data <- data.frame(y,x,z)

# deploy QAS.func-Function
result <- QAS.func(y~x+z, data=example_data,weight=NULL,seed=NULL)


}
\references{
Stoltenberg, B. (2016), Using logit on big data – from iterative methods to analytical solutions, GfK Verein Working Paper Series, No. 3 / 2016 \cr Lipovetsky, S. (2014), Analytical closed-form solution for binary logit regression by categorical predictors, Journal of Applied Statistics, No. 42 / 2015, DOI:10.1080/02664763.2014.932760
}
\author{
This method is based on the research work of Stan Lipovetsky and Birgit Stoltenberg.
}
